Moved `influxql/select` to `utils/influxql`
