+++ ui/src/dashboards/components/CellEditorOverlay.js
@@ -9,7 +9,7 @@ import Visualization from 'src/data_explorer/components/Visualization'
 import OverlayControls from 'src/dashboards/components/OverlayControls'
 import * as queryModifiers from 'src/utils/queryTransitions'
 
-import {buildSelectStatement} from 'src/data_explorer/utils/influxql/select'
+import {buildSelectStatement} from 'utils/influxql'
 
 import defaultQueryConfig from 'src/utils/defaultQueryConfig'
